fix(moe): allocate FromLogits expert_weights as bf16 regardless of logits dtype
The fp4 routing kernel writes bf16 expert weights; inheriting
routing_logits.dtype (fp32 DeepSeekV3 logits) mislabels the kernel-filled
buffer, so an unfinalized (do_finalize=False) read interprets bf16 bits as
fp32 garbage. Same bug as gh #3595, which was already fixed in the
canonical trtllm_fp4_block_scale_moe wrapper (core.py hardcodes bf16);
this aligns the unified runner with that contract and pins it with a
packing-contract test (both logits dtypes, no kernel launch).

Full do_finalize=False coverage in the unified fuzzer is tracked by
gh #3926.

+++ flashinfer/fused_moe/runners.py
@@ -325,8 +325,13 @@ class TrtllmFp4RoutedRunner(TunableRunner):
             topk_ids = act.hidden_states_q.new_empty(
                 (num_tokens, routing.top_k), dtype=torch.int32
             )
-            # routing_logits.new_empty inherits its device + dtype (core.py:2253).
-            expert_weights = routing_logits.new_empty((num_tokens, routing.top_k))
+            # MUST be bf16 regardless of logits dtype: the fp4 routing kernel
+            # writes bf16 expert weights, so inheriting fp32 from the logits
+            # mislabels the returned buffer (gh #3595 — the canonical wrapper
+            # in core.py hardcodes bf16 for the same reason).
+            expert_weights = routing_logits.new_empty(
+                (num_tokens, routing.top_k), dtype=torch.bfloat16
+            )
         elif routing_input_mode == RoutingInputMode.PackedPrecomputed:
             # Pre-routed: pack the host selection into (GLOBAL expert_id << 16) | bf16(weight).
             # The kernel expects GLOBAL ids and filters/maps them via the separately
